requirements.txt
# STG Community Website Python Dependencies
# Generated for deployment

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.33.0
gunicorn>=23.0.0
starlette>=0.44.0
itsdangerous>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.14.0

# Authentication & Security
python-jose[cryptography]>=3.4.0
# 使用 passlib + 旧版 bcrypt，避免新版本对 >72 字节密码直接抛异常
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.20

# Email
redmail>=0.6.0
aiosmtplib>=3.0.2
email-validator>=2.2.0

# Template Engine
jinja2>=3.1.6
mako>=1.3.0

# Content Processing
markdown>=3.7.0
markdown-it-py>=3.0.0
bleach>=6.1.0

# HTTP Client
httpx>=0.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Other dependencies
anyio>=4.5.0
click>=8.1.0
typer>=0.16.0
watchfiles>=0.24.0

